xvdom API update: createDynamic now accepts isOnlyChild boolean
For JSX where a dynamic is the only child of an element…

```jsx
<div>{message}</div>
```

… `xvdom.createDynamic` can optimize for this very common case in a
variety of ways.  For example, if the dynamic is an array:

1) don’t need an invisible marker node (text node) to track the end of
a rendered array
2) appendChild can be more used frequently (instead of the slower
insertBefore)
3) when rerendering an empty array, very fast `textContent = ‘’` can be
used instead of individually removing previous array elements, since
there is no risk of removing anything before or after the array

Other optimizations for text dynamics are possible (`textContent =
value` instead of creating a `textNode`).